+++ src/cmem/ActivityControl/ActivityControlTypes.ts
@@ -28,6 +28,8 @@ export interface IActivityStatus {
     runtime?: number;
     // The start time as date time, e.g. "2021-09-07T09:34:53.153Z"
     startTime?: string;
+    // The queue time spent waiting before workflow is executed as date time, e.g. "2021-09-07T09:34:53.153Z"
+    queueTime?: string;
 }
 export type SilkActivityStatusProps = IActivityStatus;