+++ src/components/Skeleton/classnames.ts
@@ -1,0 +1,3 @@
+import { Classes as BlueprintClasses } from "@blueprintjs/core";
+
+export const SKELETON = BlueprintClasses.SKELETON;