+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Fixed
 
 -   `<Pagination />`
-    -   adjust color of arrown in disabled navigation button
+    -   adjust color of arrow in disabled navigation button
 
 ## [23.3.1] - 2023-11-15