+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+-   `<PropertyValuePair />`, `<PropertyName />`, `<PropertyValue />`
+    -   `nowrap`: force display on one line without breaks
+
 ### Fixed
 
 -   `<Pagination />`