+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 
+### Added
+
+-   `<PropertyValuePair />`, `<PropertyName />`, `<PropertyValue />`
+    -   `nowrap`: force display on one line without breaks
+
+### Fixed
+
+-   `<Pagination />`
+    -   adjust color of arrow in disabled navigation button
+
 ## [23.3.1] - 2023-11-15
 
 ### Fixed