+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
     -   provides a loading state display of its children elements
 -   `<TableCell />`
     -   `alignHorizontal`: allow to center cell contents
+-   `<ActivityControlWidget />` added extra line to show timer for execution period
 
 ### Fixed